move files around regarding to https://marmelab.com/blog/2015/12/17/react-directory-structure.html issue #2
