+++ voter-front/src/app/App.css
@@ -16,12 +16,6 @@
   }
 }
 
-/* .pageContainerNarrowBox{
-  border: solid 1px grey;
-  padding: 50px;
-  margin-top: 50px;
-} */
-
 .addProposalButton{
   margin-left: 12px;
 }
