+++ voter-front/src/app/actionTypes.js
