+++ voter-front/src/app/reducers.js
