+++ voter-front/src/app/store.js
@@ -2,8 +2,8 @@ import { createStore, applyMiddleware, compose } from 'redux'
 import { routerMiddleware } from 'react-router-redux'
 import thunk from 'redux-thunk'
 import createHistory from 'history/createBrowserHistory'
-import rootReducer from './modules'
-import {getAvailableTileImages} from './modules/tileImages'
+import rootReducer from './reducers'
+import {getAvailableTileImages} from '../utils/tileImages'
 
 export const history = createHistory()
