--- voter-front/src/containers/about/index.js
@@ -1,8 +1,0 @@
-import React from 'react'
-
-export default () => (
-  <div>
-    <h1>About Us</h1>
-    <p>Hello Medium!</p>
-  </div>
-)
