+++ voter-front/src/index.js
@@ -1,18 +1,9 @@
-// import React from 'react';
-// import ReactDOM from 'react-dom';
-// import './index.css';
-// import App from './App';
-// import registerServiceWorker from './registerServiceWorker';
-
-// ReactDOM.render(<App />, document.getElementById('root'));
-// registerServiceWorker();
-
 import React from 'react'
 import { render } from 'react-dom'
 import { Provider } from 'react-redux'
 import { ConnectedRouter } from 'react-router-redux'
-import store, { history } from './store'
-import App from './containers/App.js'
+import store, { history } from './app/store'
+import App from './app/AppContainer.js'
 import registerServiceWorker from './registerServiceWorker';
