+++ voter-front/src/register/RegisterForm.js
@@ -1,23 +1,23 @@
 import React from 'react'
 import { Field, reduxForm } from 'redux-form'
-import {renderTextField,  required,minLength2, maxLength100, } from '../../modules/formUtil'
+import {renderTextField,  required,minLength2, maxLength100, validateTheSamePasswd, isEmail } from '../utils/formUtil'
 
 // Material ui
 import RaisedButton from 'material-ui/RaisedButton';
 
 
-let LogInForm = (props) => { 
+let RegisterForm = (props) => { 
     
     const { handleSubmit, pristine, submitting, reset } = props
     return <div>
                 <form onSubmit={handleSubmit}>
-                    <h1>Log In</h1>
+                    <h1>Register user</h1>
                     <Field
                         name="login"
                         component={renderTextField}
                         label="Login"
-                        validate={[required,minLength2, maxLength100]}
-                        autoFocus 
+                        validate={[required,minLength2, maxLength100, isEmail]}
+                        autoFocus
                     />
                     <br />
                     <Field
@@ -27,11 +27,19 @@ let LogInForm = (props) => {
                         type="password"
                         validate={[required,minLength2, maxLength100]}
                     />
+                     <br />
+                    <Field
+                        name="repeatPasswd"
+                        component={renderTextField}
+                        label="Reapeat password"
+                        type="password"
+                        validate={[required,minLength2, maxLength100]}
+                    />
                     <br /><br />
-                    <RaisedButton type="submit" className="logInButton" label="Log In" primary={true} disabled={pristine || submitting}  />
+                    <RaisedButton type="submit" className="regiserInButton" label="Register" primary={true} disabled={pristine || submitting}  />
                     <RaisedButton className="clearButton" label="Clear values" onClick={reset} />
                 </form>
             </div>
   };
 
-export default reduxForm({form: 'LogInForm'})(LogInForm) 
+export default reduxForm({form: 'RegisterForm', validate: validateTheSamePasswd})(RegisterForm)
