+++ voter-front/src/utils/tileImages.js
